refactor(portal/text): rename text subsystem files to match Portal vocabulary
FreeTypeContext -> TypeFaceFoundry
TextLayout      -> TypeSetter
RenderText      -> InkPress

Aligns Portal::Text naming with the craft register established in
Portal::Graphics (ShaderFoundry, TypePress, SamplerForge, RenderFlow).
